add RemoteAgent (#357)
* add RemoteAgent

Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>

Implements an agent type, `RemoteAgent` that allows for running agents
that run remotely and are accessible via a URL.

To use, declare `.spec.framework: remote` and `.spec.url: <some-url>`
in the agent YAML file.

Note that as of now, the other `.spec` values still need to be
provided but are not used by the RemoteAgent.

* updates

Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>

* update prompt emoji

Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>

---------

Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>
Co-authored-by: dr.max <maxim@us.ibm.com>